+++ src/OmnichannelChatSDK.ts
@@ -1290,7 +1290,7 @@ class OmnichannelChatSDK {
         return message;
     }
 
-    public async sendMessage(message: ChatSDKMessage): Promise<void> {
+    public async sendMessage(message: ChatSDKMessage): Promise<OmnichannelMessage | void> {
         this.scenarioMarker.startScenario(TelemetryEvent.SendMessages, {
             RequestId: this.requestId,
             ChatId: this.chatToken.chatId as string
@@ -1318,19 +1318,26 @@ class OmnichannelChatSDK {
             }
 
             try {
-                await (this.conversation as ACSConversation)?.sendMessage(sendMessageRequest);
+                const chatMessage = await (this.conversation as ACSConversation)?.sendMessage(sendMessageRequest);
 
                 this.scenarioMarker.completeScenario(TelemetryEvent.SendMessages, {
                     RequestId: this.requestId,
                     ChatId: this.chatToken.chatId as string
                 });
+
+                return chatMessage;
             } catch (error) {
+                const exceptionDetails: ChatSDKExceptionDetails = {
+                    response: ChatSDKErrorName.ChatSDKSendMessageFailed,
+                    errorObject: `${error}`
+                };
                 this.scenarioMarker.failScenario(TelemetryEvent.SendMessages, {
                     RequestId: this.requestId,
                     ChatId: this.chatToken.chatId as string,
+                    ExceptionDetails: JSON.stringify(exceptionDetails)
                 });
 
-                throw new Error('ChatSDKSendMessageFailed');
+                throw new Error(ChatSDKErrorName.ChatSDKSendMessageFailed);
             }
         } else {
             const messageToSend: IRawMessage = {
@@ -1363,13 +1370,18 @@ class OmnichannelChatSDK {
                     RequestId: this.requestId,
                     ChatId: this.chatToken.chatId as string
                 });
-            } catch {
+            } catch (error) {
+                const exceptionDetails: ChatSDKExceptionDetails = {
+                    response: ChatSDKErrorName.ChatSDKSendMessageFailed,
+                    errorObject: `${error}`
+                };
                 this.scenarioMarker.failScenario(TelemetryEvent.SendMessages, {
                     RequestId: this.requestId,
-                    ChatId: this.chatToken.chatId as string
+                    ChatId: this.chatToken.chatId as string,
+                    ExceptionDetails: JSON.stringify(exceptionDetails)
                 });
 
-                throw new Error('ChatSDKSendMessageFailed');
+                throw new Error(ChatSDKErrorName.ChatSDKSendMessageFailed);
             }
         }
     }